Add FutureBuilder to RecipeDetail
